更改代码框架，分离各个calss
