+++ main.py
@@ -1,101 +1,16 @@
 import os
 import json
 import numpy as np
-from typing import List
-
-from sqlalchemy import create_engine, Column, Integer, String, Text, DateTime, ForeignKey
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
 import astrbot.api.message_components as Comp
 from astrbot.api import AstrBotConfig, logger
 from astrbot.api.event import filter, AstrMessageEvent, MessageEventResult
 from astrbot.api.star import Context, Star, register
 from astrbot.api.message_components import Plain
+from .cluster_manager import ClusterCenter,ChatHistory,ClusterManager
+from .database import Database
+from .embedding_api import EmbeddingProvider
+from .utils import *
 
-
-Base = declarative_base()
-
-def cal_similarity(v1, v2):
-    return np.dot(v1, v2) / (np.linalg.norm(v1) * np.linalg.norm(v2))
-
-class ClusterCenter(Base):
-    __tablename__ = 'cluster_centers'
-    id = Column(Integer, primary_key=True)
-    initial_embedding = Column(Text)
-    center_embedding = Column(Text)  # 中心向量
-    member_count = Column(Integer)   # 成员数量
-
-
-class ChatHistory(Base):
-    __tablename__ = 'chat_history'
-    id = Column(Integer, primary_key=True)
-    cluster_id = Column(Integer, ForeignKey('cluster_centers.id'))  # 所属簇ID
-    message_id = Column(Text)
-    embedding = Column(Text)  # 存储JSON格式的embedding
-
-
-class ClusterManager:
-    def __init__(self, session, config):
-        self.session = session
-        self.config = config
-
-    async def find_nearest_cluster(self, embedding: List[float]):
-        # 获取所有簇中心
-        clusters = self.session.query(ClusterCenter).all()
-        if not clusters:
-            return None
-
-        # 计算与各簇中心的相似度
-        query_vec = np.array(embedding)
-        max_similarity = -1
-        best_cluster = None
-        for cluster in clusters:
-            center_vec = np.array(json.loads(cluster.center_embedding))
-            similarity = cal_similarity(query_vec, center_vec)
-            if similarity > max_similarity:
-                max_similarity = similarity
-                best_cluster = cluster
-
-        return best_cluster if max_similarity > self.config["cluster_threshold"] else None
-
-    async def update_cluster_center(self, cluster: ClusterCenter, new_embedding: List[float]):
-        # 增量更新簇中心（加权平均）
-        old_center = np.array(json.loads(cluster.center_embedding))
-        new_center = (old_center * cluster.member_count + np.array(new_embedding)) / (cluster.member_count + 1)
-        cluster.center_embedding = json.dumps(new_center.tolist())
-        cluster.member_count += 1
-
-
-        initial_center=json.loads(cluster.initial_embedding)
-        if cal_similarity(np.array(initial_center), new_center)<self.config["cluster_threshold"]:
-
-            logger.info("test")
-            cluster1= await self.new_cluster(initial_center)
-            cluster2= await self.new_cluster(new_center.tolist())
-
-            records = self.session.query(ChatHistory).filter_by(cluster_id=cluster.id).all()
-            for record in records:
-                vec = np.array(json.loads(record.embedding))
-                similarity1 = cal_similarity(np.array(initial_center), vec)
-                similarity2 = cal_similarity(new_center, vec)
-                if similarity1 > similarity2:
-                    record.cluster_id=cluster1.id
-                    await self.update_cluster_center(cluster1,json.loads(record.embedding))
-                else:
-                    record.cluster_id = cluster2.id
-                    await self.update_cluster_center(cluster2, json.loads(record.embedding))
-            self.session.delete(cluster)
-
-
-    async def new_cluster(self,new_embedding: List[float]):
-        new_cluster = ClusterCenter(
-            initial_embedding = json.dumps(new_embedding),
-            center_embedding = json.dumps(new_embedding),
-            member_count = 1
-        )
-        self.session.add(new_cluster)
-        self.session.commit()
-        return  new_cluster
 
 
 
@@ -104,58 +19,16 @@ class QQArchaeology(Star):
     def __init__(self, context: Context, config: AstrBotConfig):
         super().__init__(context)
         self.config = config
-        self._db_path=os.path.join("data","plugins","astrbot_plugin_cyber_archaeology","db")
-        self.sessions ={}
+        self.database = Database(config)
+        self.embeddingProvider = EmbeddingProvider(config)
 
 
-    def _init_db(self, unified_msg_origin: str):
-        # 按群号创建独立数据库文件
-        db_path = os.path.join(self._db_path, f"{unified_msg_origin}.db")
-        os.makedirs(os.path.dirname(db_path), exist_ok=True)
-
-        # 创建新引擎并初始化表结构[1,2](@ref)
-        engine = create_engine(f'sqlite:///{db_path}')
-        if not os.path.exists(db_path):
-            Base.metadata.create_all(engine)
-        return engine
-
-    def get_session(self, unified_msg_origin: str):
-        """获取指定群组的数据库会话"""
-        if unified_msg_origin not in self.sessions:
-            engine = self._init_db(unified_msg_origin)
-            Session = sessionmaker(bind=engine)
-            self.sessions[unified_msg_origin] = Session()
-        return self.sessions[unified_msg_origin]
-
-    async def get_embedding(self, text: str) -> List[float]:
-        """调用Ollama获取embedding（异步版本）"""
-        try:
-            import httpx
-            async with httpx.AsyncClient(timeout=30) as client:
-                resp = await client.post(
-                    f"{self.config['ollama_api_url']}/api/embeddings",
-                    json={
-                        "model": self.config["embed_model"],
-                        "prompt": text
-                    }
-                )
-                print(resp)
-                resp.raise_for_status()  # 自动处理4xx/5xx状态码
-                return resp.json()["embedding"]
-        except httpx.HTTPStatusError as e:
-            logger.error(f"API错误: {e.response.status_code} - {e.response.text}")
-        except httpx.RequestError as e:
-            logger.error(f"网络请求失败: {str(e)}")
-        except json.JSONDecodeError:
-            logger.error("响应数据解析失败")
-        except Exception as e:
-            logger.error(f"未知错误: {str(e)}")
 
     @filter.command("search",alias={'考古'})
     async def search_command(self, event: AstrMessageEvent, query: str):
         """搜索历史记录 示例：/search 关键词"""
         unified_msg_origin = event.unified_msg_origin
-        session = self.get_session(unified_msg_origin)  # 获取当前群的会话
+        session = self.database.get_session(unified_msg_origin)  # 获取当前群的会话
         group_id=event.get_group_id()
 
         if not query:
@@ -163,7 +36,7 @@ class QQArchaeology(Star):
             return
 
         # 获取查询embedding
-        query_embedding = await self.get_embedding(query)
+        query_embedding = await self.embeddingProvider.get_embedding(query)
         if not query_embedding:
             yield event.plain_result("Embedding服务不可用")
             return
@@ -179,7 +52,7 @@ class QQArchaeology(Star):
         for cluster in clusters:
             center_vec = np.array(json.loads(cluster.center_embedding))
             similarity = cal_similarity(query_vec, center_vec)
-            if similarity > self.config["cluster_threshold"]:
+            if similarity > self.config["plugin_conf"]["cluster_threshold"]:
                 candidate_clusters.append(cluster)
 
         # 第二阶段：在候选簇内精确搜索
@@ -189,12 +62,12 @@ class QQArchaeology(Star):
             for record in records:
                 db_vec = np.array(json.loads(record.embedding))
                 similarity = cal_similarity(query_vec, db_vec)
-                if similarity > self.config["threshold"]:
+                if similarity > self.config["plugin_conf"]["threshold"]:
                     results.append((similarity, record))
 
         # 排序并取前K个
         results.sort(reverse=True, key=lambda x: x[0])
-        top_results = results[:self.config["top_k"]]
+        top_results = results[:self.config["plugin_conf"]["top_k"]]
 
         # 构造返回结果
         if not top_results:
@@ -228,35 +101,6 @@ class QQArchaeology(Star):
 
         event.stop_event()
 
-            # 构造获取群消息历史的请求参数
-            # payloads = {
-            #     "message_id": rec[1].message_id,
-            # }
-            #
-            # # 调用API获取群聊历史消息
-            # msg = await client.api.call_action("get_msg", **payloads)
-            #
-            # # 处理消息历史记录，对其格式化
-            # message_text = ""
-            # messagechain = msg.get("message", [])
-            # for part in messagechain:
-            #     if part['type'] == 'text':
-            #         message_text += part['data']['text'].strip() + " "
-            #     elif part['type'] == 'json':  # 处理JSON格式的分享卡片等特殊消息
-            #         try:
-            #             json_content = json.loads(part['data']['data'])
-            #             if 'desc' in json_content.get('meta', {}).get('news', {}):
-            #                 message_text += f"[分享内容]{json_content['meta']['news']['desc']} "
-            #         except:
-            #             pass
-            #
-            #     # 表情消息处理
-            #     elif part['type'] == 'face':
-            #         message_text += "[表情] "
-            #
-            #     # 生成标准化的消息记录格式
-            # if message_text:
-            #     yield event.plain_result(message_text)
 
 
     @filter.event_message_type(filter.EventMessageType.GROUP_MESSAGE)
@@ -264,7 +108,7 @@ class QQArchaeology(Star):
         """保存群聊历史记录"""
         try:
             unified_msg_origin = event.unified_msg_origin
-            session = self.get_session(unified_msg_origin)  # 获取对应群组的会话
+            session = self.database.get_session(unified_msg_origin)  # 获取对应群组的会话
 
             # 获取消息文本
             messagechain = event.message_obj.message
@@ -277,7 +121,7 @@ class QQArchaeology(Star):
                 return
 
             # 生成embedding
-            embedding = await self.get_embedding(message)
+            embedding = await self.embeddingProvider.get_embedding(message)
 
             if not embedding:
                 return
@@ -306,13 +150,13 @@ class QQArchaeology(Star):
         except Exception as e:
             logger.error(f"保存记录失败: {str(e)}")
             unified_msg_origin = event.unified_msg_origin
-            self.get_session(unified_msg_origin).rollback()
+            self.database.get_session(unified_msg_origin).rollback()
 
     async def terminate(self):
         """关闭所有数据库连接"""
-        for unified_msg_origin, session in self.sessions.items():
+        for unified_msg_origin, session in self.database.sessions.items():
             session.close()
-        self.sessions.clear()
+        self.database.sessions.clear()
 
     @filter.command_group("cyber_archaeology",alias={'ca'})
     def cyber_archaeology(self):
@@ -326,9 +170,9 @@ class QQArchaeology(Star):
         import shutil
         try:
             # 关闭所有现存会话
-            for session in self.sessions.values():
+            for session in self.database.sessions.values():
                 session.close()
-            self.sessions.clear()
+            self.database.sessions.clear()
 
             # 删除整个数据库目录
             if os.path.exists(self._db_path):
@@ -348,7 +192,7 @@ class QQArchaeology(Star):
     async def clear_current_command(self, event: AstrMessageEvent):
         """清空当前群聊记录 示例：/ca clear"""
         try:
-            session = self.get_session(event.unified_msg_origin)
+            session = self.database.get_session(event.unified_msg_origin)
 
             # 删除当前群所有记录
             session.query(ChatHistory).delete()
@@ -358,7 +202,7 @@ class QQArchaeology(Star):
         except Exception as e:
             logger.error(f"清空本群记录失败: {str(e)}")
             unified_msg_origin = event.unified_msg_origin
-            self.get_session(unified_msg_origin).rollback()
+            self.database.get_session(unified_msg_origin).rollback()
             yield event.plain_result("清空操作失败，请检查日志")
 
     @filter.permission_type(filter.PermissionType.ADMIN)
@@ -366,7 +210,7 @@ class QQArchaeology(Star):
     async def load_history_command(self, event: AstrMessageEvent, count: int = None, seq: int = 0):
         """读取插件未安装前bot所保存的历史数据当前群聊记录 示例：/ca load_history <读取消息条数:int> [初始消息序号:int]"""
         try:
-            session = self.get_session(event.unified_msg_origin)
+            session = self.database.get_session(event.unified_msg_origin)
 
             from astrbot.core.platform.sources.aiocqhttp.aiocqhttp_message_event import AiocqhttpMessageEvent
             assert isinstance(event, AiocqhttpMessageEvent)
@@ -415,7 +259,7 @@ class QQArchaeology(Star):
 
 
                 # 获取embedding
-                embedding = await self.get_embedding(message_text)
+                embedding = await self.embeddingProvider.get_embedding(message_text)
 
                 if not embedding:
                     return
@@ -450,7 +294,7 @@ class QQArchaeology(Star):
         except Exception as e:
             logger.error(f"导入本群记录失败: {str(e)}")
             unified_msg_origin = event.unified_msg_origin
-            self.get_session(unified_msg_origin).rollback()
+            self.database.get_session(unified_msg_origin).rollback()
             yield event.plain_result("导入本群记录失败，请检查日志")
