+++ utils.py
@@ -1,0 +1,4 @@
+import numpy as np
+
+def cal_similarity(v1, v2):
+    return np.dot(v1, v2) / (np.linalg.norm(v1) * np.linalg.norm(v2))
